Get a single user moved to service
