+++ app/src/main/java/com/example/test/myreceipts/BLL/Callback.java
@@ -1,0 +1,12 @@
+package com.example.test.myreceipts.BLL;
+
+import com.example.test.myreceipts.Entity.User;
+
+/**
+ * Created by thomas on 24-04-2018.
+ */
+
+public interface Callback {
+
+    void act(User model);
+}
